# specs/001-project-management-oa/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 项目管理OA系统 API
  description: 基于Excel需求的项目管理OA系统API文档
  version: 1.0.0
  contact:
    name: 开发团队
    email: dev@company.com
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.project-oa.com/api/v1
    description: 生产环境
security:
  - BearerAuth: []
paths:
...
  /projects/{id}/business:
    get:
      tags:
        - 项目经营信息
      summary: 获取项目经营信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 经营信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectBusiness'
        '404':
          description: 项目不存在

    put:
      tags:
        - 项目经营信息
      summary: 更新项目经营信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectBusinessRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectBusiness'
        '404':
          description: 项目不存在

  # 专家费支付管理
  /projects/{id}/expert-fee-payments:
    get:
      tags:
        - 项目经营信息
      summary: 获取项目专家费支付列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 专家费支付列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpertFeePayment'
        '404':
          description: 项目不存在

    post:
      tags:
        - 项目经营信息
      summary: 添加专家费支付记录
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpertFeePaymentRequest'
      responses:
        '201':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertFeePayment'
        '404':
          description: 项目不存在

  # 项目成员管理
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        real_name:
          type: string
        role:
          type: string
          enum: [admin, manager, business, designer, reviewer, finance]
        department:
          type: string
        phone:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProjectBusiness:
      type: object
      description: 项目经营信息
      properties:
        project_id:
          type: integer
        client_id:
          type: integer
          nullable: true
          description: 甲方ID（可选，最多关联一个）
        client:
          $ref: '#/components/schemas/Client'
          nullable: true
          description: 甲方信息（可选）
        contact_name:
          type: string
          description: 联系人姓名（可选）
        contact_phone:
          type: string
          description: 联系电话（可选）
        business_manager:
          type: string
          description: 经营负责人（可选）
        bidding_info:
          type: object
          description: 招投标信息（可选）
          properties:
            bidding_documents:
              type: array
              items:
                $ref: '#/components/schemas/File'
            bidding_status:
              type: string
              enum: [not_started, in_progress, won, lost]
        contracts:
          type: array
          items:
            $ref: '#/components/schemas/Contract'
        updated_at:
          type: string
          format: date-time
    UpdateProjectBusinessRequest:
      type: object
      description: 更新项目经营信息请求
      properties:
        client_id:
          type: integer
          nullable: true
          description: 甲方ID（可选，设置为null可删除关联）
        contact_name:
          type: string
          description: 联系人姓名（可选）
        contact_phone:
          type: string
          description: 联系电话（可选）
        business_manager:
          type: string
          description: 经营负责人（可选）
    Client:
      type: object
      properties:
        id:
          type: integer
        client_name:
          type: string
          description: 甲方名称（唯一，2-100字符）
          minLength: 2
          maxLength: 100
        contact_name:
          type: string
          description: 联系人姓名（可选）
        contact_phone:
          type: string
          description: 联系电话（可选，11位手机号或固定电话格式）
        email:
          type: string
          format: email
          description: 邮箱地址（可选，有效邮箱格式）
        address:
          type: string
          description: 地址（可选）
        tax_number:
          type: string
          description: 统一社会信用代码（可选，18位格式）
        bank_account:
          type: string
          description: 银行账号（可选）
        bank_name:
          type: string
          description: 开户银行（可选）
        is_active:
          type: boolean
          description: 是否启用（默认true）
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Contract:
      type: object
      properties:
        id:
          type: integer
        contract_number:
          type: string
        contract_type:
          type: string
        sign_date:
          type: string
          format: date
        contract_rate:
          type: number
          format: float
        contract_amount:
          type: number
          format: float
          description: 合同金额（应等于设计费+勘察费+咨询费之和）
        design_fee:
          type: number
          format: float
          description: 设计费
        survey_fee:
          type: number
          format: float
          description: 勘察费
        consultation_fee:
          type: number
          format: float
          description: 咨询费
        file_path:
          type: string
        project_id:
          type: integer
        amendments:
          type: array
          items:
            $ref: '#/components/schemas/ContractAmendment'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ContractAmendment:
      type: object
      properties:
        id:
          type: integer
        amendment_number:
          type: string
        sign_date:
          type: string
          format: date
        file_path:
          type: string
        description:
          type: string
        contract_id:
          type: integer
        contract:
          $ref: '#/components/schemas/Contract'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    File:
      type: object
      properties:
        id:
          type: integer
        file_name:
          type: string
        original_name:
          type: string
        file_path:
          type: string
        file_size:
          type: integer
        file_type:
          type: string
        mime_type:
          type: string
        category:
          type: string
          enum: [contract, bidding, design, audit, production, other]
        description:
          type: string
        project_id:
          type: integer
        uploader_id:
          type: integer
        uploader:
          $ref: '#/components/schemas/User'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ExpertFeePayment:
      type: object
      properties:
        id:
          type: integer
        payment_method:
          type: string
          enum: [cash, transfer]
          description: 支付方式（现金/转账）
        amount:
          type: number
          format: float
          description: 金额
        expert_name:
          type: string
          description: 专家姓名
        expert_id:
          type: integer
          nullable: true
          description: 专家用户ID（如果是系统内用户）
        expert:
          $ref: '#/components/schemas/User'
        description:
          type: string
        project_id:
          type: integer
        created_by_id:
          type: integer
        created_by:
          $ref: '#/components/schemas/User'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateExpertFeePaymentRequest:
      type: object
      required:
        - payment_method
        - amount
        - expert_name
      properties:
        payment_method:
          type: string
          enum: [cash, transfer]
          description: 支付方式（现金/转账）
        amount:
          type: number
          format: float
          description: 金额
        expert_name:
          type: string
          description: 专家姓名
        expert_id:
          type: integer
          nullable: true
          description: 专家用户ID（如果是系统内用户，可选）
        description:
          type: string
...